Extended new structure to penalized projection algorithm
