+++ applications/ShapeOptimizationApplication/python_scripts/analyzer_base.py
@@ -1,0 +1,27 @@
+# ==============================================================================
+#  KratosShapeOptimizationApplication
+#
+#  License:         BSD License
+#                   license: ShapeOptimizationApplication/license.txt
+#
+#  Main authors:    Baumgärtner Daniel, https://github.com/dbaumgaertner
+#
+# ==============================================================================
+
+# Making KratosMultiphysics backward compatible with python 2.6 and 2.7
+from __future__ import print_function, absolute_import, division 
+
+# importing the Kratos Library
+from KratosMultiphysics import *
+from KratosMultiphysics.ShapeOptimizationApplication import *
+
+# check that KratosMultiphysics was imported in the main script
+CheckForPreviousImport()
+
+# ==============================================================================
+class analyzerBaseClass:
+
+    def analyzeDesignAndReportToCommunicator( self, currentDesign, optimizationIteration, communicator ):
+        raise RuntimeError("Analyzer base class is called. Please check your implementation of the function >> analyzeDesignAndReportToCommunicator << .")
+
+# ==============================================================================
