+++ applications/ShapeOptimizationApplication/test_examples/01_Strain_Energy_Minimization_3D_Hook/Optimization_Results/response_log.csv
@@ -1,0 +1,12 @@
+itr	,	f	,	df_absolute[%]	,	df_relative[%]	,	step_size[-]	,	t_iteration[s]	,	t_total[s],	time_stamp
+1	,	3.890143572846	,	0.00	,	0.000000	,	2.75	,	11.7	,	11.7	,	Thu Apr  6 08:21:12 2017
+2	,	3.778653268796	,	-2.87	,	-2.865969	,	2.75	,	13.8	,	25.5	,	Thu Apr  6 08:21:25 2017
+3	,	3.669572796020	,	-5.67	,	-2.804022	,	2.75	,	11.7	,	37.2	,	Thu Apr  6 08:21:37 2017
+4	,	3.563183103386	,	-8.40	,	-2.734853	,	2.75	,	11.2	,	48.4	,	Thu Apr  6 08:21:48 2017
+5	,	3.460083934511	,	-11.06	,	-2.650266	,	2.75	,	12.1	,	60.5	,	Thu Apr  6 08:22:00 2017
+6	,	3.360824036782	,	-13.61	,	-2.551574	,	2.75	,	12.9	,	73.4	,	Thu Apr  6 08:22:13 2017
+7	,	3.265882903732	,	-16.05	,	-2.440556	,	2.75	,	11.6	,	85.0	,	Thu Apr  6 08:22:25 2017
+8	,	3.175395337224	,	-18.37	,	-2.326073	,	2.75	,	11.1	,	96.2	,	Thu Apr  6 08:22:36 2017
+9	,	3.089247162934	,	-20.59	,	-2.214524	,	2.75	,	11.9	,	108.0	,	Thu Apr  6 08:22:48 2017
+10	,	3.007512285250	,	-22.69	,	-2.101076	,	2.75	,	12.5	,	120.5	,	Thu Apr  6 08:23:00 2017
+11	,	2.929799919073	,	-24.69	,	-1.997673	,	2.75	,	11.2	,	131.7	,	Thu Apr  6 08:23:11 2017
