+++ applications/ShapeOptimizationApplication/test_examples/01_Strain_Energy_Minimization_3D_Hook/ProjectParameters_with_mass_constraint.json
@@ -166,20 +166,20 @@
             "step_size"                  : 3
         },
         "output" : {
-            "output_directory"          : "Optimization_Results_With_Mass_Constraint",
-            "design_history_filename"   : "design_history",
-            "optimization_log_filename" : "optimization_log",
-            "nodal_results"             : [ "NORMALIZED_SURFACE_NORMAL",
-                                            "OBJECTIVE_SENSITIVITY",
-                                            "CONSTRAINT_SENSITIVITY",
-                                            "MAPPED_OBJECTIVE_SENSITIVITY",
-                                            "MAPPED_CONSTRAINT_SENSITIVITY",
-                                            "SEARCH_DIRECTION",
-                                            "DESIGN_UPDATE",
-                                            "SHAPE_UPDATE",
-                                            "SHAPE_CHANGE_ABSOLUTE" ],            
-            "output_type" : {
-                "format"                 : "GiD",
+            "output_directory"        : "Optimization_Results_With_Mass_Constraint",
+            "design_history_filename" : "design_history",
+            "response_log_filename"   : "response_log",
+            "nodal_results"           : [ "NORMALIZED_SURFACE_NORMAL",
+                                          "OBJECTIVE_SENSITIVITY",
+                                          "CONSTRAINT_SENSITIVITY",
+                                          "MAPPED_OBJECTIVE_SENSITIVITY",
+                                          "MAPPED_CONSTRAINT_SENSITIVITY",
+                                          "SEARCH_DIRECTION",
+                                          "DESIGN_UPDATE",
+                                          "SHAPE_UPDATE",
+                                          "SHAPE_CHANGE_ABSOLUTE" ],            
+            "output_format" : {
+                "name"                   : "GiD",
                 "VolumeOutput"           : true,
                 "GiDPostMode"            : "Binary",
                 "GiDWriteMeshFlag"       : true,
