+++ applications/ShapeOptimizationApplication/test_examples/02_Strain_Energy_Minimization_3D_Shell/3D_Shell_fine.mdpa
@@ -78338,7 +78338,7 @@ Begin SubModelPart PointLoad3D_load // Group load // Subtree PointLoad3D
     End SubModelPartConditions
 End SubModelPart
 
-Begin SubModelPart design_surface_fine
+Begin SubModelPart design_surface
     Begin SubModelPartNodes
         1
         2
