+++ applications/ShapeOptimizationApplication/test_examples/02_Strain_Energy_Minimization_3D_Shell/Optimization_Results_Fine/response_log.csv
@@ -1,0 +1,8 @@
+itr	,	f	,	df_absolute[%]	,	df_relative[%]	,	step_size[-]	,	t_iteration[s]	,	t_total[s],	time_stamp
+1	,	4.281807967403	,	0.00	,	0.000000	,	0.1	,	19.9	,	19.9	,	Thu Apr  6 08:16:45 2017
+2	,	4.147360527201	,	-3.14	,	-3.139969	,	0.1	,	19.7	,	39.5	,	Thu Apr  6 08:17:04 2017
+3	,	3.970893635130	,	-7.26	,	-4.121317	,	0.1	,	19.5	,	59.0	,	Thu Apr  6 08:17:24 2017
+4	,	3.701449347087	,	-13.55	,	-6.292769	,	0.1	,	20.1	,	79.2	,	Thu Apr  6 08:17:44 2017
+5	,	3.406532083879	,	-20.44	,	-6.887681	,	0.1	,	20.2	,	99.4	,	Thu Apr  6 08:18:04 2017
+6	,	3.113135667121	,	-27.29	,	-6.852162	,	0.1	,	20.2	,	119.7	,	Thu Apr  6 08:18:25 2017
+7	,	2.840793971756	,	-33.65	,	-6.360437	,	0.1	,	20.3	,	139.9	,	Thu Apr  6 08:18:45 2017
