+++ applications/ShapeOptimizationApplication/test_examples/02_Strain_Energy_Minimization_3D_Shell/ProjectParameters_fine_mesh.json
@@ -117,7 +117,7 @@
         "design_variables" : {
             "variable_type"             : "vertex_morphing", 
             "input_model_part_name"     : "3D_Shell_fine",
-            "design_submodel_part_name" : "design_surface_fine",
+            "design_submodel_part_name" : "design_surface",
             "domain_size"               : 3,
             "filter" : {
                 "filter_function_type" : "linear",
@@ -153,18 +153,18 @@
             "step_size"                  : 0.1
         },
         "output" : {
-            "output_directory"          : "Optimization_Results_Fine_Mesh",
-            "design_history_filename"   : "design_history",
-            "optimization_log_filename" : "optimization_log",
-            "nodal_results"             : [ "NORMALIZED_SURFACE_NORMAL",
-                                            "OBJECTIVE_SENSITIVITY",
-                                            "MAPPED_OBJECTIVE_SENSITIVITY",
-                                            "DESIGN_UPDATE",
-                                            "DESIGN_CHANGE_ABSOLUTE",
-                                            "SHAPE_UPDATE",
-                                            "SHAPE_CHANGE_ABSOLUTE" ],            
-            "output_type" : {
-                "format"                 : "GiD",
+            "output_directory"        : "Optimization_Results_Fine",
+            "design_history_filename" : "design_history",
+            "response_log_filename"   : "response_log",
+            "nodal_results"           : [ "NORMALIZED_SURFACE_NORMAL",
+                                          "OBJECTIVE_SENSITIVITY",
+                                          "MAPPED_OBJECTIVE_SENSITIVITY",
+                                          "DESIGN_UPDATE",
+                                          "DESIGN_CHANGE_ABSOLUTE",
+                                          "SHAPE_UPDATE",
+                                          "SHAPE_CHANGE_ABSOLUTE" ],            
+            "output_format" : {
+                "name"                   : "gid",
                 "VolumeOutput"           : true,
                 "GiDPostMode"            : "Binary",
                 "GiDWriteMeshFlag"       : true,
@@ -172,5 +172,5 @@
                 "GiDMultiFileFlag"       : "Single"
             }      
         }                  
-    }    
+    }
 }
