+++ applications/ShapeOptimizationApplication/test_examples/02_Strain_Energy_Minimization_3D_Shell/run_optimization.py
@@ -58,7 +58,9 @@ for i in range(ProjectParameters["solver_settings"]["processes_sub_model_part_li
 # Analyzer
 # ======================================================================================================================================
 
-class kratosCSMAnalyzer( optimizerFactory.analyzerBaseClass ):
+class kratosCSMAnalyzer( (__import__("analyzer_base")).analyzerBaseClass ):
+    
+    # --------------------------------------------------------------------------    
     def __init__( self ):
 
         self.initializeGIDOutput()
